backend/requirements.txt
blinker==1.9.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
python-dotenv==1.0.1
Werkzeug==3.1.3
Flask-Cors==5.0.0
zipp==3.21.0
importlib_metadata==8.5.0